requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==4.2.29
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
ndg-httpsclient==0.5.1
numpy==2.1.3
pillow==12.1.1
psycopg2-binary==2.9.10
pusher==3.3.2
PyYAML==6.0.1
pyasn1==0.6.3
pycparser==2.23
PyJWT==2.12.0
PyNaCl==1.6.2
pyOpenSSL==26.0.0
python-dateutil==2.9.0
python-dotenv==1.0.1
RapidFuzz==3.14.3
requests==2.32.5
scipy==1.16.3
sentry-sdk==2.12.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
uWSGI==2.0.28
whitenoise==6.7.0
